requirements.txt
python-telegram-bot==20.*
requests==2.*
python-dotenv==1.*
langchain==0.1.*
langchain-openai==0.0.*
langgraph==0.0.*
openai==1.*
pydantic==2.*
